+++ src/client/app/routes/admin/configuredrule.js
@@ -39,7 +39,8 @@ export default Ember.Route.extend(RouteMixin, RulesetEmberizer, {
 				return { result: result, meta: meta};
 			}),
 			importers: this.store.findAll('importer'),
-			exporters: this.store.findAll('exporter')
+			exporters: this.store.findAll('exporter'),
+			authGroups: this.store.findAll('authgroup')
 
 		})
 	},